auth-service/app/requirements.txt
requests
pyjwt[crypto]
diskcache
